remove ineffective handshaking
It seems that mimicking the the communication that's seen in the logic
analyser trace is not enough; it appears that the handset needs to
respond to communication from the controller and does not with the set
up that exists, so remove it and directly wire the controller's RX to
the handset's TX with a wire passthrough. This does not fix the issue,
but reduces the amount of work that we were doing and appears to allow
the system to re-handshake after a single handset button press and wait
until the handset display blanks. The RX from the controller is retained
so that the MitM is still aware of the desk height.